[1986-01-16] RZ 1.14 / SZ 1.14 (suspect) - Add CRC-32 error detection
Chuck's original USENET posting of this release could not be fully recovered
and no original release archive files have been located. So the code in this
commit (notably sz.c) is likely a fork of an earlier release that was
merged with some of Chuck's official changes (adding 32-bit CRC support) for
version 1.14. See CHANGELOG.md for details.