Fix warnings regarding usage of utime()
Was missing required prototype (in utime.h) and was passing a different
(but compatible) argument type. I don't see any reaosn for timep to be global,
but I left it so.

Though POSIX.1-2008 marks utime() as obsolete (utimes apparently is now
preferred), to minimize the diff, I left it as-is/was.  Changing to use
utimes() instead would also be trivial:

  #include <sys/time.h> // utimes()

  struct timeval timep[2];

  timep[0].tv_sec = time(NULL);
  timep[1].tv_sec = Modtime;
  utimes(Pathname, timep);